refactor(worktree-doctor): split into types/parse/classify/execute submodules
Brings packages/crane-mcp/src/tools/worktree-doctor.ts under structural
caps (was 600 LOC).

- worktree-doctor.ts (barrel, 52 LOC): re-exports public symbols.
- worktree-doctor/types.ts: input schema + interfaces + PidChecker type.
- worktree-doctor/parse.ts: resolveWorktreesDir, parseWorktreeList,
  classifyLock. parseWorktreeList complexity reduced from 18 to <15 via
  extracted applyPorcelainLine helper.
- worktree-doctor/classify.ts: classifyWorktree + 6 gate helpers
  (gateOwnWorktree, gateLock, gateLsof, gateFreshHead, gateCleanTree,
  gateMerged) — each under 75 LOC.
- worktree-doctor/execute.ts: executeWorktreeDoctor split into
  scanCandidates, fetchWorktreeIndex, applyRemove, processCandidate,
  fetchMergedPrs. max-params fixed via 2-param struct instead of 8 flat.

Public API + signatures preserved. 21/21 tests pass.
